cathay/requirements.txt
requests
beautifulsoup4
mysql.connector
loguru
